Refactor

+++ Web Quiz Engine/task/src/engine/controllers/QuizController.java
@@ -1,7 +1,7 @@
 package engine.controllers;
 
 import engine.models.Quiz;
-import engine.QuizService;
+import engine.services.QuizService;
 import engine.models.CompletionDto;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
@@ -18,7 +18,6 @@ import java.util.NoSuchElementException;
 
 import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
 
-//@CrossOrigin(origins = "*", maxAge = 3600)
 @RestController
 public class QuizController {
     private static final String SERVICE_WARNING_MESSAGE = "An error has occurred";
@@ -28,23 +27,12 @@ public class QuizController {
     public QuizController() {
     }
 
-//    @GetMapping("/api/quiz")
-//    public ResponseEntity<String> getQuiz() {
-//        return ResponseEntity.ok()
-//                .header(HttpHeaders.CONTENT_TYPE, String.valueOf(MediaType.APPLICATION_JSON))
-//                .body("{\n" +
-//                        "  \"title\": \"The Java Logo\",\n" +
-//                        "  \"text\": \"What is depicted on the Java logo?\",\n" +
-//                        "  \"options\": [\"Robot\",\"Tea leaf\",\"Cup of coffee\",\"Bug\"]\n" +
-//                        "}\n");
-//    }
     @Secured("ROLE_USER")
-//    @PreAuthorize("hasRole('USER')")
     @GetMapping("/api/quizzes/{id}")
-    public ResponseEntity<String> getQuizById(@PathVariable int id) {
+    public ResponseEntity<String> getQuizById(@PathVariable long id) {
         Quiz quiz;
         try {
-            quiz = quizService.getQuizById((long) (id));
+            quiz = quizService.getQuizById((id));
         } catch (Exception e) {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND, SERVICE_WARNING_MESSAGE);
         }
@@ -66,18 +54,6 @@ public class QuizController {
             @RequestParam(defaultValue = "10") Integer pageSize,
             @RequestParam(defaultValue = "id") String sortBy)
     {
-//        List<Quiz> list = quizService.getAllQuizzes(page, pageSize, sortBy);
-//        String result = null;
-//        try {
-//            result = new ObjectMapper().writeValueAsString(list);
-//        } catch (JsonProcessingException e) {
-//            e.printStackTrace();
-//        }
-//        final HttpHeaders httpHeaders = new HttpHeaders();
-//        httpHeaders.setContentType(MediaType.APPLICATION_JSON);
-//        if (list.isEmpty()){
-//            result="{\"content\":[]}";
-//        }
         return quizService.getAllQuizzes(page, pageSize, sortBy);
     }
 
@@ -92,13 +68,6 @@ public class QuizController {
     {
         return quizService.getAllCompletions(page, pageSize, sortBy,principal);
     }
-
-//    @PostMapping(value = "/api/quiz",consumes = "application/json")
-//    ResponseEntity<String> answerQuiz(@RequestBody Integer[] answer) {
-//        return ResponseEntity.ok()
-//                .header(HttpHeaders.CONTENT_TYPE, String.valueOf(MediaType.APPLICATION_JSON))
-//                .body(Quiz.saveAnswer(answer));
-//    }
 
     @Secured("ROLE_USER")
     @PostMapping(value = "/api/quizzes", consumes = "application/json")
@@ -126,7 +95,6 @@ public class QuizController {
                                      @RequestBody Quiz answer,
                                      @Autowired Principal principal) {
         try {
-//            Quiz quiz = quizService.getQuizById((long) (id));
             return ResponseEntity.ok()
                     .header(HttpHeaders.CONTENT_TYPE, String.valueOf(MediaType.APPLICATION_JSON))
                     .body(quizService.solve(id,answer.getAnswer(),principal.getName()));
